+++ src/pydvl/influence/torch/__init__.py
@@ -2,5 +2,6 @@ from .influence_function_model import (
     ArnoldiInfluence,
     CgInfluence,
     DirectInfluence,
+    EkfacInfluence,
     LissaInfluence,
 )